Former-commit-id: 28f7579c4579600104ff93f14e890a23f9968416

+++ pqueens/schedulers/local_scheduler.py
@@ -1,24 +1,19 @@
+""" There should be a docstring """
 
 import os
-import subprocess
 import sys
-import pathlib
+from .scheduler import Scheduler
 
-from pqueens.schedulers.scheduler import Scheduler
 
 class LocalScheduler(Scheduler):
     """ Scheduler which submits jobs to the local machine via a shell command"""
 
-    def __init__(self, scheduler_name):
-        """ Create LocalScheduler
-
-        Args:
-            scheduler_name (string):    Name of scheduler
-        """
+    def __init__(self, base_settings, scheduler_name):
         self.name = scheduler_name
+        super(LocalScheduler, self).__init__(base_settings)
 
     @classmethod
-    def from_config_create_scheduler(cls, scheduler_name, config):
+    def from_config_create_scheduler(cls, config, base_settings, scheduler_name=None):
         """ Create scheduler from config dictionary
 
         Args:
@@ -28,66 +23,19 @@ class LocalScheduler(Scheduler):
         Returns:
             scheduler:              Instance of LocalScheduler
         """
+        # pre assemble some strings as base_settings
+        base_settings['output'] = None
+        base_settings['tasks_info'] = None
+        base_settings['walltime_info'] = None
+        base_settings['job_flag'] = None
+        base_settings['scheduler_start'] = None
+        base_settings['command_line_opt'] = None
+        base_settings['scheduler_options'] = None
 
-        return cls(scheduler_name)
+        return cls(base_settings, scheduler_name)
 
-
-    def submit(self, job_id, experiment_name, batch, experiment_dir,
-               database_address, driver_params={}):
-        """ Submit job locally by calling subprocess
-
-        Args:
-            job_id (int):               Id of job to be started
-            experiment_name (string):   Name of experiment
-            batch (string):             Batch number
-            experiment_dir (string):    Directory to write output to
-            database_address (string):  Address of MongoDB database
-            driver_options (dict):      Options for driver (optional)
-
-        Returns:
-            int: id of process associated with the job,
-                 or None if submission failed
-
-        """
-        # TODO implement proper driver hierarchy 
-        # TODO find a better way to do this
-        #base_path = os.path.dirname(os.path.realpath(pqueens.__file__))
-        base_path = pathlib.Path(__file__).parent.parent
-
-
-        # get path to Python interpreter
-        # scheduler and driver should be called with the same Python executable
-        python_path = sys.executable
-
-        # assemble shell command
-        cmd = ('%s %s/drivers/gen_driver_local.py --db_address %s --experiment_name '
-               '%s --job_id %s --batch %s' %
-               (python_path, base_path, database_address, experiment_name, job_id, batch))
-
-
-        output_directory = os.path.join(experiment_dir, 'output')
-        if not os.path.isdir(output_directory):
-            os.mkdir(output_directory)
-
-        output_filename = os.path.join(output_directory, '%08d.out' % job_id)
-        output_file = open(output_filename, 'w')
-
-        process = subprocess.Popen(cmd,
-                                   stdout=output_file,
-                                   stderr=output_file,
-                                   shell=True)
-
-        process.poll()
-        if process.returncode is not None and process.returncode < 0:
-            sys.stderr.write("Failed to submit job or job crashed "
-                             "with return code %d !\n" % process.returncode)
-            return None
-        else:
-            sys.stderr.write("Submitted job as process: %d\n" % process.pid)
-            return process.pid
-
-
-    def alive(self, process_id):
+# ------------------- CHILD METHODS THAT MUST BE IMPLEMENTED ------------------
+    def alive(self, process_id):  # TODO: ok for now (gets called in resources)
         """ Check whether or not job is still running
 
         Args:
@@ -96,10 +44,30 @@ class LocalScheduler(Scheduler):
         Returns:
             bool: indicator if job is still alive
         """
-        try:
-            # Send an alive signal to proc
-            os.kill(process_id, 0)
-        except OSError:
+        alive = False
+        command_list = ['ps h -p', str(process_id)]
+        command_string = ' '.join(command_list)
+        stdout, _, p = super().run_subprocess(command_string)
+
+        if stdout:
+            sys.stderr.write("Job %d waiting in queue.\n" % (process_id))
+            alive = True
+        else:
+            sys.stderr.write("Job %d is held or suspended.\n" % (process_id))
+            alive = False
+
+        if not alive:
+            try:
+                # try to kill the job.
+                os.kill(process_id, 0)
+                sys.stderr.write("Killed job %d.\n" % (process_id))
+            except ValueError:
+                sys.stderr.write("Failed to kill job %d.\n" % (process_id))
+
             return False
         else:
             return True
+
+    def get_process_id_from_output(self):
+        """ docstring """
+        pass
